+++ crates/swc_ecma_lints/src/rules/no_param_reassign.rs
@@ -8,7 +8,6 @@ use swc_common::{
     Span,
 };
 use swc_ecma_ast::*;
-use swc_ecma_utils::ident::IdentLike;
 use swc_ecma_visit::{noop_visit_type, Visit, VisitWith};
 
 use crate::{