+++ crates/swc_ecma_minifier/src/compress/optimize/collapse_vars.rs
@@ -1,5 +1,4 @@
 use swc_ecma_ast::*;
-use swc_ecma_utils::ident::IdentLike;
 
 use super::Optimizer;
 use crate::mode::Mode;