+++ crates/swc_ecma_minifier/src/compress/optimize/evaluate.rs
@@ -3,7 +3,7 @@ use std::num::FpCategory;
 use swc_atoms::js_word;
 use swc_common::{util::take::Take, Spanned, SyntaxContext, DUMMY_SP};
 use swc_ecma_ast::*;
-use swc_ecma_utils::{ident::IdentLike, undefined, ExprExt, Value::Known};
+use swc_ecma_utils::{undefined, ExprExt, Value::Known};
 
 use super::Optimizer;
 use crate::{compress::util::eval_as_number, mode::Mode, DISABLE_BUGGY_PASSES};
@@ -85,7 +85,11 @@ where
             return;
         }
 
-        if self.ctx.is_delete_arg || self.ctx.is_update_arg || self.ctx.is_lhs_of_assign {
+        if self.ctx.is_delete_arg
+            || self.ctx.is_update_arg
+            || self.ctx.is_lhs_of_assign
+            || self.ctx.in_with_stmt
+        {
             return;
         }