+++ crates/swc_ecma_minifier/src/compress/optimize/iife.rs
@@ -5,7 +5,7 @@ use swc_atoms::js_word;
 use swc_common::{pass::Either, util::take::Take, Mark, Spanned, SyntaxContext, DUMMY_SP};
 use swc_ecma_ast::*;
 use swc_ecma_utils::{
-    contains_arguments, contains_this_expr, find_ids, ident::IdentLike, undefined, ExprFactory, Id,
+    contains_arguments, contains_this_expr, find_ids, undefined, ExprFactory, Id,
 };
 use swc_ecma_visit::VisitMutWith;