+++ crates/swc_ecma_minifier/src/compress/optimize/strings.rs
@@ -1,7 +1,7 @@
 use swc_atoms::js_word;
 use swc_common::{util::take::Take, Spanned, SyntaxContext};
 use swc_ecma_ast::*;
-use swc_ecma_utils::{ident::IdentLike, ExprExt, Value::Known};
+use swc_ecma_utils::{ExprExt, Value::Known};
 
 use super::Optimizer;
 use crate::mode::Mode;