+++ crates/swc_ecma_minifier/src/eval.rs
@@ -5,7 +5,7 @@ use swc_atoms::js_word;
 use swc_common::{collections::AHashMap, util::take::Take, DUMMY_SP};
 use swc_ecma_ast::*;
 use swc_ecma_transforms_optimization::simplify::{expr_simplifier, ExprSimplifierConfig};
-use swc_ecma_utils::{ident::IdentLike, undefined, ExprExt, Id};
+use swc_ecma_utils::{undefined, ExprExt, Id};
 use swc_ecma_visit::{FoldWith, VisitMutWith};
 
 use crate::{