+++ crates/swc_ecma_transforms_module/src/util.rs
@@ -16,8 +16,8 @@ use swc_common::{
 };
 use swc_ecma_ast::*;
 use swc_ecma_utils::{
-    ident::IdentLike, member_expr, private_ident, quote_ident, quote_str, undefined,
-    DestructuringFinder, ExprFactory, Id,
+    member_expr, private_ident, quote_ident, quote_str, undefined, DestructuringFinder,
+    ExprFactory, Id,
 };
 use swc_ecma_visit::{Fold, FoldWith, VisitWith};